ttsstt/requirements.txt
Flask==2.0.2
Flask-SocketIO==5.1.1
flask-cors>=3.0.10
flask-talisman==0.8.1
flask-limiter==3.7.0
python-dotenv==0.19.1
pyaudio==0.2.14
SpeechRecognition==3.8.1
openai>=1.37.1
Werkzeug==2.0.1
redis==4.5.1
gevent==24.2.1
gevent-websocket==0.10.1
gunicorn==22.0.0